+++ test/commands/geoadd.spec.js
@@ -4,7 +4,7 @@ var config = require("../lib/config");
 var helper = require("../helper");
 var redis = config.redis;
 
-describe("The 'getoadd' method", function () {
+describe("The 'geoadd' method", function () {
 
     helper.allTests(function(parser, ip, args) {
 
@@ -19,6 +19,7 @@ describe("The 'getoadd' method", function () {
             });
 
             it('returns 1 if the key exists', function (done) {
+                helper.serverVersionAtLeast.call(this, client, [3, 2, 0]);
                 client.geoadd("mycity:21:0:location", "13.361389","38.115556","COR", function(err, res) {
                     console.log(err, res);
                     // geoadd is still in the unstable branch. As soon as it reaches the stable one, activate this test