+++ changelog.md
@@ -14,6 +14,7 @@ Bugfixes:
 -  Fix multi.hmset key not being type converted if used with an object and key not being a string (@BridgeAR)
 -  Fix parser errors not being catched properly (@BridgeAR)
 -  Fix a crash that could occur if a redis server does return the info command as usual #541 (@BridgeAR)
+-  Explicitly passing undefined as a callback statement will work again. E.g. client.publish('channel', 'message', undefined); (@BridgeAR)
 
 ## v2.0.1 - Sep 24, 2015